refactor: jpa erd data type Long to Integer - id값을 long으로 하고 있었는데, erd에 맞게 integer로 변경

+++ backend/ttabong/src/main/java/com/ttabong/controller/sns/ReviewController.java
@@ -24,7 +24,7 @@ public class ReviewController {
 
     /* 후기 삭제 */
     @PatchMapping("/{reviewId}/delete")
-    public ResponseEntity<ReviewDeleteResponseDto> deleteReview(@PathVariable Long reviewId) {
+    public ResponseEntity<ReviewDeleteResponseDto> deleteReview(@PathVariable Integer reviewId) {
         ReviewDeleteResponseDto response = reviewService.deleteReview(reviewId);
         return ResponseEntity.ok(response);
     }
