+++ backend/ttabong/src/main/java/com/ttabong/repository/recruit/RecruitRepository.java
@@ -3,5 +3,5 @@ package com.ttabong.repository.recruit;
 import com.ttabong.entity.recruit.Recruit;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface RecruitRepository extends JpaRepository<Recruit, Long> {
+public interface RecruitRepository extends JpaRepository<Recruit, Integer> {
 }
