+++ backend/ttabong/src/main/java/com/ttabong/service/sns/ReviewService.java
@@ -6,5 +6,5 @@ import com.ttabong.dto.sns.response.ReviewDeleteResponseDto;
 
 public interface ReviewService {
     ReviewCreateResponseDto createReview(ReviewCreateRequestDto reviewCreateRequestDto);
-    ReviewDeleteResponseDto deleteReview(Long reviewId);
+    ReviewDeleteResponseDto deleteReview(Integer reviewId);
 }
